[Feature] Write a universal motor controller interface
In 2025, we need to be able to control both a Spark and TalonFX powered drive-train. Since we had to re-write the Spark MAX driver anyway, implement both Spark MAX and Talon FX using the same interface.

There are plenty of missing features compared to the previous abstraction, but this should be enough for a swerve chassis.